+++ Pokemon 2048/Tile.h
@@ -10,14 +10,14 @@ class Tile {
 		void ResetSize();
 		void FlagMerged(bool flag);
 		bool HasMerged();
-		void SetOffset(float x, float y);
+		void SetOffset(int x, int y);
 		void Draw(Vector2 screenDimensions, int x, int y);
 
 	private:
 		int state;
 		bool merged = false;
-		float tileSizePercentage = 0.1;
-		float tileOffsetPercentage = 1;
+		double tileSizePercentage = 0.1;
+		double tileOffsetPercentage = 1;
 		Vector2 tileShiftOffset = { 0, 0 };
 		Color colours[10] = {RED, ORANGE, YELLOW, GREEN, SKYBLUE, DARKBLUE, DARKPURPLE, PURPLE, MAGENTA, PINK};
 };
